requirements.txt
streamlit
google-genai
pypdf
pillow